requirements.txt
flask
requests
upstox-python-sdk
websocket-client
python-binance
python-dotenv
pandas
numpy
scipy
apscheduler

